+++ src/histogram/histogram.cpp
@@ -10,6 +10,8 @@
 #include <boost/histogram/python/storage.hpp>
 #include <boost/histogram/python/histogram.hpp>
 #include <boost/histogram/python/histogram_fill.hpp>
+#include <boost/histogram/python/histogram_atomic.hpp>
+#include <boost/histogram/python/histogram_threaded.hpp>
 
 #include <boost/histogram.hpp>
 #include <boost/histogram/axis/ostream.hpp>
@@ -96,6 +98,12 @@ py::class_<bh::histogram<A, S>> register_histogram_by_type(py::module& m, const 
       [](histogram_t &self, py::args args) {
         boost::mp11::mp_with_index<BOOST_HISTOGRAM_DETAIL_AXES_LIMIT>(args.size(), fill_helper<histogram_t>(self, args));
       }, "Insert data into histogram")
+    
+    // generic threaded fill for 1 to N args
+    .def("threaded_fill",
+         [](histogram_t &self, unsigned threads, py::args args) {
+             boost::mp11::mp_with_index<BOOST_HISTOGRAM_DETAIL_AXES_LIMIT>(args.size(), fill_helper_threaded<histogram_t>(self, args, threads));
+         }, "threads"_a, "Insert data into histogram in threads (0 for machine cores)")
 
     .def("at", [](histogram_t &self, py::args &args) {
         // Optimize for no dynamic?
@@ -141,15 +149,18 @@ py::class_<bh::histogram<A, S>> register_histogram_by_type(py::module& m, const 
 }
 
 template<typename A, typename S>
-void add_mt_fill(py::class_<bh::histogram<A, S>>& hist) {
+void add_atomic_fill(py::class_<bh::histogram<A, S>>& hist) {
     using histogram_t = bh::histogram<A, S>;
     hist.def("mtfill", [](histogram_t &self, size_t threads, py::args args) {
-        boost::mp11::mp_with_index<BOOST_HISTOGRAM_DETAIL_AXES_LIMIT>(args.size(), fill_helper_mt<histogram_t>(self, args, threads));
+        boost::mp11::mp_with_index<BOOST_HISTOGRAM_DETAIL_AXES_LIMIT>(args.size(), fill_helper_atomic<histogram_t>(self, args, threads));
     }, "threads"_a, "Insert data into histogram, in some number of threads (0 for default)")
     ;
 }
 
 void register_histogram(py::module& m) {
+    
+    m.attr("BOOST_HISTOGRAM_DETAIL_AXES_LIMIT") = BOOST_HISTOGRAM_DETAIL_AXES_LIMIT;
+    
     py::module hist = m.def_submodule("hist");
 
     // Fast specializations: Fixed number of axis (may be removed if above versions are fast enough)
@@ -167,7 +178,7 @@ void register_histogram(py::module& m) {
     auto regular_atomic_int_1d = register_histogram_by_type<axes::regular_1D, storage::atomic_int>(hist,
         "regular_atomic_int_1d",
         "1-dimensional histogram for int valued data (atomic).");
-    add_mt_fill(regular_atomic_int_1d);
+    add_atomic_fill(regular_atomic_int_1d);
     
     m.def("make_histogram", [](axis::regular_uoflow& ax1, storage::atomic_int){
         return bh::make_histogram_with(storage::atomic_int(), ax1);
@@ -213,7 +224,7 @@ void register_histogram(py::module& m) {
     auto regular_atomic_int = register_histogram_by_type<axes::regular, storage::atomic_int>(hist,
         "regular_atomic_int",
         "N-dimensional histogram for atomic int-valued data.");
-    add_mt_fill(regular_atomic_int);
+    add_atomic_fill(regular_atomic_int);
 
     register_histogram_by_type<axes::regular_noflow, storage::int_>(hist,
         "regular_noflow_int",
@@ -228,7 +239,7 @@ void register_histogram(py::module& m) {
     auto any_atomic_int = register_histogram_by_type<axes::any, storage::atomic_int>(hist,
         "any_atomic_int",
         "N-dimensional histogram for int-valued data with any axis types (threadsafe).");
-    add_mt_fill(any_atomic_int);
+    add_atomic_fill(any_atomic_int);
 
     register_histogram_by_type<axes::any, storage::double_>(hist,
         "any_double",
